pyproject.toml
[project]
name = "thesis-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rfdetr>=1.4.3",
    "supervision>=0.27.0.post1",
]
